requirements.txt
alembic==1.12.1
bcrypt==3.1.7
boto3==1.14.2
botocore==1.17.2
cffi==1.16.0
click==8.1.7
docutils==0.15.2
Flask==2.1.3
Flask-Bcrypt==0.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gunicorn==19.9.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.0.0
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.21
python-dateutil==2.8.2
s3transfer==0.3.7
six==1.16.0
SQLAlchemy==1.3.17
timeago==1.0.14
typing_extensions==4.8.0
urllib3==1.25.11
Werkzeug==2.0.3
WTForms==3.1.0
zipp==3.17.0
